refactor: centralize storage mock fixture
