+++ tests/mocks/storage/__init__.py
@@ -1,5 +1,10 @@
 """Mock implementations for storage testing."""
 
 from .adapters import MockGitAdapter, MockStorageCoordinator
+from .fixtures import storage_mock
 
-__all__ = ['MockGitAdapter', 'MockStorageCoordinator'] 
+__all__ = [
+    'MockGitAdapter',
+    'MockStorageCoordinator',
+    'storage_mock'
+]
